hw3/pyproject.toml
[tool.poetry]
name = "mlops-hw1"
version = "0.1.0"
description = ""
authors = ["Ilya Popov <ilpopov@ozon.ru>"]
readme = "README.md"
packages = [{include = "web_app"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "2.1.2"
flask-restx = "^0.5.1"
Werkzeug = "2.1.2"
pandas = "^1.5.0"
sklearn = "^0.0"
xgboost = "^1.6.2"
grpcio = "^1.49.1"
grpcio-tools = "^1.49.1"
scikit-learn = "^1.1.3"
pyyaml = "^6.0"
psycopg2 = "^2.9.5"
sqlalchemy = "^1.4.44"
pytest = "^7.2.0"
pytest-mock = "^3.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
